revert SOVERSION back to 0 and library version to 0.0.0
while 0.0.0 is probably not exactly right, this is what was
shipped in unshield up to 1.2. the change to cmake in 1.3
made the version change to ${PROJECT_VERSION}, which seems wrong,
as library versions usually don't relate to the project version.

there were no symbol changes between 1.0 and 1.3.
1.4 only introduced a new symbol (unshield_is_unicode) which does
not warrant to bump the soname.

fixes #41